# openapi/openapi.gen.yaml
openapi: 3.1.0
info:
  title: Loomtale Studio API
  version: 0.1.0
  description: Internal story-to-YouTube video studio API.
servers:
  - url: /api/v1
paths:
...
  /series/{id}/bible:
    get:
      operationId: getBible
      summary: The series' story bible sections
      tags:
        - series
      x-min-role: viewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: story bible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryBible'
        '404':
          description: series or bible not found in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      operationId: updateBibleSection
      summary: Update one bible section with an optimistic-concurrency check
      tags:
        - series
      x-min-role: editor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BibleSectionUpdateRequest'
      responses:
        '200':
          description: bible section updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryBible'
        '409':
          description: expectedVersion did not match the section's current version
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /episodes:
    get:
      operationId: listEpisodes
      summary: Cursor-paginated list of a series' episodes, with aggregated draft status (no N+1)
      tags:
        - episodes
      x-min-role: viewer
      parameters:
        - name: seriesId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: page of episodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodeList'
    post:
      operationId: createEpisode
      summary: Append a new, empty episode to a series
      tags:
        - episodes
      x-min-role: editor
      parameters:
        - name: seriesId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: episode created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
  /episodes/{id}:
    get:
      operationId: getEpisode
      summary: Episode outline and metadata
      tags:
        - episodes
      x-min-role: viewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: episode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        '404':
          description: episode not found in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      operationId: updateEpisode
      summary: Update episode title/status
      tags:
        - episodes
      x-min-role: editor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EpisodeUpdateRequest'
      responses:
        '200':
          description: episode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        '404':
          description: episode not found in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /episodes/{id}/drafts/{lang}:
    get:
      operationId: getDraft
      summary: The episode's draft for one language
      tags:
        - episodes
      x-min-role: viewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lang
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TargetLanguage'
      responses:
        '200':
          description: draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodeDraft'
        '404':
          description: episode or draft not found in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      operationId: patchDraft
      summary: Autosave paragraph ops with an optimistic-concurrency version check
      tags:
        - episodes
      x-min-role: editor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lang
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TargetLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftPatchRequest'
      responses:
        '200':
          description: draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodeDraft'
        '409':
          description: expectedVersion is stale; client should refetch and merge
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    TargetLanguage:
      type: string
      enum:
        - en
        - vi
...
    Origin:
      type: string
      enum:
        - user
        - import
        - model
    BibleSection:
      type: object
      required:
        - content
        - origin
        - tainted
        - version
      properties:
        content:
          type: string
          description: Plain text for prose sections; JSON-encoded array text for the glossary section.
        origin:
          $ref: '#/components/schemas/Origin'
        tainted:
          type: boolean
        version:
          type: integer
    StoryBible:
      type: object
      required:
        - seriesId
        - sections
        - updatedAt
      properties:
        seriesId:
          type: string
          format: uuid
        sections:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/BibleSection'
        updatedAt:
          type: string
          format: date-time
    BibleSectionUpdateRequest:
      type: object
      required:
        - section
        - content
        - expectedVersion
      properties:
        section:
          type: string
          enum:
            - world
            - cultivation_realms
            - arcs
            - style_guide
            - running_summary
            - glossary
        content:
          type: string
          maxLength: 20000
        expectedVersion:
          type: integer
          description: Optimistic-concurrency check against the section's current version; a mismatch returns 409.
    OutlineBeat:
      type: object
      required:
        - id
        - summary
        - targetWords
      properties:
        id:
          type: string
        summary:
          type: string
        targetWords:
          type: integer
    DraftStatus:
      type: object
      properties:
        wordCount:
          type: integer
        version:
          type: integer
    Episode:
      type: object
      required:
        - id
        - seriesId
        - idx
        - title
        - outline
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
        idx:
          type: integer
        title:
          type: string
        outline:
          type: array
          items:
            $ref: '#/components/schemas/OutlineBeat'
        status:
          type: string
          enum:
            - planned
            - outlined
            - drafting
            - draft
            - reviewed
        drafts:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DraftStatus'
        durationEstimateMinutes:
          type: object
          additionalProperties:
            type: number
        createdAt:
          type: string
          format: date-time
    EpisodeList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
        nextCursor:
          type: string
    EpisodeUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        status:
          type: string
          enum:
            - planned
            - outlined
            - drafting
            - draft
            - reviewed
...
    DraftParagraph:
      type: object
      required:
        - id
        - text
        - origin
        - tainted
      properties:
        id:
          type: string
        text:
          type: string
        origin:
          $ref: '#/components/schemas/Origin'
        tainted:
          type: boolean
    EpisodeDraft:
      type: object
      required:
        - episodeId
        - lang
        - paragraphs
        - version
        - wordCount
      properties:
        episodeId:
          type: string
          format: uuid
        lang:
          $ref: '#/components/schemas/TargetLanguage'
        paragraphs:
          type: array
          items:
            $ref: '#/components/schemas/DraftParagraph'
        version:
          type: integer
        wordCount:
          type: integer
        durationEstimateMinutes:
          type: number
        summary:
          type: string
        summaryTainted:
          type: boolean
    ParagraphOp:
      type: object
      required:
        - op
        - paragraphId
      properties:
        op:
          type: string
          enum:
            - upsert
            - delete
            - move
        paragraphId:
          type: string
        text:
          type: string
        afterParagraphId:
          type: string
          description: For op=move, the paragraph id to place this one after (empty string means first).
    DraftPatchRequest:
      type: object
      required:
        - expectedVersion
        - ops
      properties:
        expectedVersion:
          type: integer
        ops:
          type: array
          items:
            $ref: '#/components/schemas/ParagraphOp'
          minItems: 1
          maxItems: 200
...